webvtt/api/VTTCue/position.html

<!doctype html>
<title>VTTCue.position</title>
<link rel="help" href="https://w3c.github.io/webvtt/#dom-vttcue-position">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
test(function(){
    var cue = new VTTCue(0, 1, 'text');

    cue.position = 'auto';
    assert_equals(cue.position, 'auto');

    for (i = 0; i <= 100; i++) {
        cue.position = i;
        assert_equals(cue.position, i);
    }

    [-1, -100, -101, 101, 200, 201].forEach(function(invalid) {
        assert_throws('IndexSizeError', function() {
            cue.position = invalid;
        });
    });

    cue.position = 1.5;
    assert_equals(cue.position, 1.5);
}, document.title+', script-created cue');
</script>
